refactor(web): move directory of modals to `features/project-schedule`
