+++ apps/web/src/features/project-schedule/model/constants/schedule.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
-import { RepeatOption } from '~/widgets/modal/ui/create-schedule';
-import { RepeatOptionItem } from '~/widgets/modal/types';
+import { RepeatOption } from '~/features/project-schedule/ui';
+import { RepeatOptionItem } from '~/features/project-schedule/types';
 
 export const REPEAT_OPTION = {
   week: { title: '매주', element: <RepeatOption.Week /> },
