+++ apps/web/src/features/project-schedule/ui/index.ts
@@ -2,3 +2,5 @@ export { default as EntireSchedule } from './EntireSchedule';
 export { default as EntireTodo } from './EntireTodo';
 export { default as ScheduleCalendar } from './ScheduleCalendar';
 export { default as ScheduleItem } from './ScheduleItem';
+export { default as RepeatOption } from './RepeatOption';
+export { default as CreateScheduleModal } from './CreateScheduleModal';
