--- apps/web/src/widgets/modal/ui/create-schedule/index.ts
@@ -1,2 +1,0 @@
-export { default as RepeatOption } from './RepeatOption';
-export { default as CreateScheduleModal } from './CreateScheduleModal';
